// frontend/src/component/TimeTable/types/class-period-option-datas.ts
export type ClassPeriodOptionDatas = {
  kamoku_name: string;
  kamokuid: number;
  kamoku_resume: string;
  kamoku_class: string;
  kamoku_day: string;
  kamoku_time: string;
  kamoku_unit: number;
  kamoku_department: string;
  kamoku_season: string;
  kamoku_num: number;
  kamoku_teacher: string;
  kamoku_status: string;
};